将pid类型改成long，简化了文件相关的代码逻辑

+++ src/main/java/com/jdb/recovery/LogRecord.java
@@ -34,7 +34,7 @@ public abstract class LogRecord {
 
     public long getPrevLsn(){return -1L;} ;
 
-    public abstract int getPageId();
+    public abstract long getPageId();
 
     public long getLsn(){return lsn;};
